requirements.txt
google-adk
google-generativeai
pinecone
python-dotenv
groq
fastapi
uvicorn
python-multipart